+++ src/containers/SidePanelDetails/Skills/index.js
@@ -9,22 +9,29 @@ import {
   getSelectedEntity,
   isEntityFetching
 } from '../../../redux/modules/entities/selectors';
-import {
-  setSelectedSubEntityId,
-  toggleListItemEntity
-} from '../../../redux/modules/entities';
-import {
-  getDependantEntityTableItems,
-  getSidePanelTableItems
-} from '../../../redux/modules/entities/listItemSelectors';
+import { setSelectedSubEntityId, toggleListItemEntity, updateProficiency } from '../../../redux/modules/entities';
+import { getSidePanelTableItems } from '../../../redux/modules/entities/listItemSelectors';
+import { getSkillMemberSidePanelTableItems } from '../../../redux/modules/entities/skills/selectors';
 import { entitiesMetaData } from '../../../redux/modules/entities/metaData';
+import store from '../../../redux/store';
 
-export function mapStateToProps(state, props) {
+export function mapStateToProps(state) {
   return {
     item: getSelectedEntity(state).toJS(),
     userHasUpdatePermission: userHasUpdatePermission(state),
-    usersItems: getDependantEntityTableItems(state),
-    usersFields: entitiesMetaData.users.memberListTableFields,
+    usersItems: getSkillMemberSidePanelTableItems(state),
+    usersFields: [
+      ...entitiesMetaData.users.memberListTableFields,
+      {
+        label: 'Proficiency',
+        name: 'proficiency',
+        editable: userHasUpdatePermission(state),
+        filterable: false,
+        actions: {
+          onChange: (id, newValue) => store.dispatch(updateProficiency(id, newValue))
+        }
+      }
+    ],
     usersFetching: isEntityFetching(state, 'users'),
     outboundIdentifierListsItems: getSidePanelTableItems(state, 'outboundIdentifierLists'),
     outboundIdentifierListsFields: entitiesMetaData.outboundIdentifierLists.memberListTableFields,