+++ src/redux/modules/entities/skills/selectors.js
@@ -1,7 +1,16 @@
 /*
  * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
  */
-import { getCurrentForm } from '../../form/selectors';
 
-export const getHasProficiencyFormValue = state =>
-  getCurrentForm(state) && getCurrentForm(state).getIn(['values', 'hasProficiency']);
+import { getSelectedEntityId } from '../selectors';
+import { getDependantEntityTableItems } from '../listItemSelectors';
+
+export const getSkillMemberSidePanelTableItems = state => {
+  const selectedEntityId = getSelectedEntityId(state);
+  return getDependantEntityTableItems(state).map(user => ({
+    ...user,
+    firstName: user.firstName !== null ? user.firstName : '',
+    lastName: user.lastName !== null ? user.lastName : '',
+    proficiency: user.skills.filter(skill => selectedEntityId === skill.id).map(skill => skill.proficiency)[0]
+  }));
+};